+++ appFirewall/C workers/log.h
@@ -33,8 +33,8 @@ typedef struct log_line_t {
 	conn_raw_t raw;
 } log_line_t;
 
-int get_log_size(void);
-log_line_t* get_log_row(int row);
+size_t get_log_size(void);
+log_line_t* get_log_row(size_t row);
 log_line_t* find_log_by_conn(char* name, conn_raw_t* c, int debug);
 //void get_log_addr_name(int row, char* str, int len);
 void append_log(char* str, char* long_str, struct bl_item_t* bl_item, conn_raw_t *raw, int blocked);
